move Gradle tasks into task package
