+++ runners/gradle-plugin/src/main/kotlin/org/jetbrains/dokka/gradle/DokkaMultiModuleFileLayout.kt
@@ -1,6 +1,7 @@
 package org.jetbrains.dokka.gradle
 
 import org.jetbrains.dokka.DokkaException
+import org.jetbrains.dokka.gradle.tasks.*
 import java.io.File
 
 /**
