+++ runners/gradle-plugin/src/main/kotlin/org/jetbrains/dokka/gradle/DokkaPlugin.kt
@@ -6,6 +6,7 @@ import org.gradle.api.Project
 import org.gradle.api.artifacts.Dependency
 import org.gradle.kotlin.dsl.register
 import org.gradle.util.GradleVersion
+import org.jetbrains.dokka.gradle.tasks.*
 
 open class DokkaPlugin : Plugin<Project> {
     override fun apply(project: Project) {
