+++ runners/gradle-plugin/src/main/kotlin/org/jetbrains/dokka/gradle/GradleDokkaSourceSetBuilderFactory.kt
@@ -1,6 +1,7 @@
 package org.jetbrains.dokka.gradle
 
 import org.gradle.api.NamedDomainObjectFactory
+import org.jetbrains.dokka.gradle.tasks.*
 
 @Suppress("ObjectLiteralToLambda") // Will fail at runtime in Gradle versions <= 6.6
 fun AbstractDokkaTask.gradleDokkaSourceSetBuilderFactory(): NamedDomainObjectFactory<GradleDokkaSourceSetBuilder> =
