+++ runners/gradle-plugin/src/main/kotlin/org/jetbrains/dokka/gradle/tasks/AbstractDokkaLeafTask.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.dokka.gradle
+package org.jetbrains.dokka.gradle.tasks
 
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.internal.plugins.DslObject
@@ -6,6 +6,9 @@ import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.Nested
 import org.gradle.kotlin.dsl.container
 import org.gradle.work.DisableCachingByDefault
+import org.jetbrains.dokka.gradle.*
+import org.jetbrains.dokka.gradle.checkSourceSetDependencies
+import org.jetbrains.dokka.gradle.kotlinOrNull
 
 @DisableCachingByDefault(because = "Abstract super-class, not to be instantiated directly")
 abstract class AbstractDokkaLeafTask : AbstractDokkaTask() {
