+++ runners/gradle-plugin/src/main/kotlin/org/jetbrains/dokka/gradle/tasks/DokkaTask.kt
@@ -1,8 +1,10 @@
-package org.jetbrains.dokka.gradle
+package org.jetbrains.dokka.gradle.tasks
 
 import org.jetbrains.dokka.DokkaConfigurationImpl
 import org.jetbrains.dokka.build
 import org.gradle.api.tasks.*
+import org.jetbrains.dokka.gradle.getSafe
+import org.jetbrains.dokka.gradle.getValidVersionOrNull
 
 @CacheableTask
 abstract class DokkaTask : AbstractDokkaLeafTask() {
